+++ src/projects/detail/components/Accordion/Accordion.jsx
@@ -30,7 +30,8 @@ const TYPE = {
   SKILLS: 'skills',
   SLIDER_RADIO: 'slide-radiogroup',
   SLIDER_STANDARD: 'slider-standard',
-  SELECT_DROPDOWN: 'select-dropdown'
+  SELECT_DROPDOWN: 'select-dropdown',
+  TALENT_PICKER: 'talent-picker',
 }
 
 /**
@@ -168,6 +169,10 @@ class Accordion extends React.Component {
     case TYPE.SLIDER_RADIO: return mapValue(value)
     case TYPE.SLIDER_STANDARD: return mapDecisionValue(value)
     case TYPE.SELECT_DROPDOWN: return mapValue(value)
+    case TYPE.TALENT_PICKER: {
+      const getRoleName = (role) => _.find(options, { role }).roleTitle
+      return _.filter(value, (v) => v.people !== '0').map((v) => `${getRoleName(v.role)}: ${v.people}`).join(', ')
+    }
     default: return value
     }
   }
@@ -189,7 +194,7 @@ class Accordion extends React.Component {
               {title}
             </div>
           </Tooltip>
-          <div styleName="value">{this.formatValue()}</div>
+          <div title={this.formatValue()} styleName="value">{this.formatValue()}</div>
           <div styleName="toggle">
             {isOpen ? <IconX styleName="toggle-icon" /> : <IconCarretDown styleName="toggle-icon" />}
           </div>
